requirements.txt
# Baseline V1 Live Trading Requirements

# Core dependencies
pandas>=2.0.0,<2.2.0
numpy>=1.24.0,<2.0.0
pytz>=2023.3
python-dotenv>=1.0.0

# OpenAlgo Python SDK
openalgo>=1.0.0

# Telegram notifications
requests>=2.31.0
python-telegram-bot>=20.0

# PostgreSQL support (for Railway deployment)
psycopg2-binary>=2.9.0

# Monitoring & Dashboard
streamlit>=1.28.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Utilities
python-dateutil>=2.8.2
